eliminate 64-bit warnings

+++ Core/NTApiRequestProcessor.m
@@ -115,7 +115,7 @@
     {
         NSHTTPURLResponse *httpResponse = (NSHTTPURLResponse *)response;
         
-        _response.httpStatusCode = httpResponse.statusCode;
+        _response.httpStatusCode = (int)httpResponse.statusCode;
         _response.headers = httpResponse.allHeaderFields;
     }
